add big integer support, extend initialized program size

+++ src/main/java/com/martak/adventofcode/amplificationcircuit/AmplifierFactory.java
@@ -3,16 +3,17 @@ package com.martak.adventofcode.amplificationcircuit;
 import com.martak.adventofcode.intcodecomputer.CodeComputer;
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 public class AmplifierFactory {
 
-    public static List<Amplifier> getAmplifiers(int[] phaseSequence, int[] controllerSoftware) {
+    public static List<Amplifier> getAmplifiers(BigInteger[] phaseSequence, BigInteger[] controllerSoftware) {
         List<Amplifier> amplifiers = new ArrayList<>(phaseSequence.length);
         for (int i = 0; i < 5; i++) {
-            long[] inputs = {phaseSequence[i], 0};
+            BigInteger[] inputs = {phaseSequence[i], BigInteger.valueOf(0)};
             Context context = new Context(0, Arrays.copyOf(controllerSoftware, controllerSoftware.length), inputs, "", 0);
             amplifiers.add(new Amplifier(new CodeComputer(), context));
         }
