+++ src/main/java/com/martak/adventofcode/intcodecomputer/model/argmode/RelativeArgMode.java
@@ -2,13 +2,15 @@ package com.martak.adventofcode.intcodecomputer.model.argmode;
 
 import com.martak.adventofcode.intcodecomputer.model.Context;
 
+import java.math.BigInteger;
+
 public class RelativeArgMode implements ArgMode {
 
     @Override
-    public int getValue(Context context, int pointer) {
-        int[] code = context.getCode();
+    public BigInteger getValue(Context context, int pointer) {
+        BigInteger[] code = context.getCode();
         int relativeBase = context.getRelativeBase();
-        int address = relativeBase + code[pointer];
+        int address = relativeBase + code[pointer].intValue();
         return code[address];
     }
 }
